\name{createBuffer}
\alias{createBuffer}
\title{Create a buffer...}
\usage{createBuffer(sContext, sMemFlag, sGlobalWorkSize, sObject)
}
\description{Create a buffer}
\value{pointer to the buffer}
\author{Willem Ligtenberg}
\arguments{\item{sContext}{context}
\item{sMemFlag}{memflag}
\item{sGlobalWorkSize}{globalworksize}
\item{sObject}{object}
}

