\name{createContextFromType}
\alias{createContextFromType}
\title{Creates an OpenCL context for a given device type}
\description{
   Creates an OpenCL context
}
\details{
  Calls clCreateContextFromType using specified device type.
}
\references{
\url{http://www.khronos.org/opencl/}
}
\keyword{OpenCL}