\name{createProgram}
\alias{createProgram}
\title{Create a program...}
\usage{createProgram(context, kernelSrc, kernelName, ...)
}
\description{Create a program}
\value{pointer to the kernel}
\author{Willem Ligtenberg}
\arguments{\item{context}{context in which to execute the kernel}
\item{kernelSrc}{source code of the kernel function as a string}
\item{kernelName}{name of the kernel}
\item{...}{further arguments}
}

