\name{createProgramWithSource}
\alias{createProgramWithSource}
\title{Creates an OpenCL program}
\description{
   Creates an OpenCL program
}
\details{
  Calls clCreateProgramWithSource using specified OpenCL context and source code (as String).
}
\references{
\url{http://www.khronos.org/opencl/}
}
\keyword{OpenCL}