\name{enqueueNDRangeKernel}
\alias{enqueueNDRangeKernel}
\title{Enqueues an NDRangeKernel}
\description{
   Enqueues an NDRangeKernel
}
\details{
  Calls clEnqueueNDRangeKernel using specified queue, kernel global work size and local work size.
}
\references{
\url{http://www.khronos.org/opencl/}
}
\keyword{OpenCL}