\name{enqueueReadBuffer}
\alias{enqueueReadBuffer}
\title{Enqueue the read buffer...}
\usage{enqueueReadBuffer(sQueue, sMemBuffer, sGlobalWorkSize, sObject)
}
\description{Enqueue the read buffer}
\value{pointer}
\author{Willem Ligtenberg}
\arguments{\item{sQueue}{queue}
\item{sMemBuffer}{membuffer}
\item{sGlobalWorkSize}{global worksize}
\item{sObject}{object}
}

