\name{enqueueReadBufferIntegerVector}
\alias{enqueueReadBufferIntegerVector}
\title{Enqueues a read buffer for a vector of integers.}
\description{
   Enqueues a read buffer for a vector of integers.
}
\details{
  Calls clEnqueueReadBuffer function for the object that was passed as the argument.
  This method is for internal use, users are encouraged to use enqueueReadBuffer instead.
}
\references{
\url{http://www.khronos.org/opencl/}
}
\keyword{OpenCL}