\name{enqueueWriteBuffer}
\alias{enqueueWriteBuffer}
\title{Enqueue a write buffer...}
\usage{enqueueWriteBuffer(sQueue, sMemBuffer, sGlobalWorkSize, sObject)
}
\description{Enqueue a write buffer}
\value{pointer}
\author{Willem Ligtenberg}
\arguments{\item{sQueue}{queue}
\item{sMemBuffer}{membufffer}
\item{sGlobalWorkSize}{globalworksize}
\item{sObject}{object}
}

