\name{enqueueWriteBufferFloatVector}
\alias{enqueueWriteBufferFloatVector}
\title{Enqueues a write buffer for a vector of floats.}
\description{
   Enqueues a write buffer for a vector of floats.
}
\details{
  Calls the right enqueueWriteBuffer function for the class of the object that was passed as the argument.
  This method is for internal use, users are encouraged to use enqueueWriteBuffer instead.
}
\references{
\url{http://www.khronos.org/opencl/}
}
\keyword{OpenCL}