\name{getContextInfo}
\alias{getContextInfo}
\title{Retreives information of the OpenCL context.}
\description{
   Retreives information of the OpenCL context.
}
\details{
  Calls clGetContextInfo, currently only CL_CONTEXT_NUM_DEVICES is supported.
}
\references{
\url{http://www.khronos.org/opencl/}
}
\keyword{OpenCL}