\name{getDeviceIDs}
\alias{getDeviceIDs}
\title{Retreives the different IDs of the available OpenCL devices.}
\description{
   Retreives the different IDs of the available OpenCL devices.
}
\details{
  Calls clGetDeviceIDs and returns a list of device IDs which can be used later on.
}
\references{
\url{http://www.khronos.org/opencl/}
}
\keyword{OpenCL}