\name{getPlatformIDs}
\alias{getPlatformIDs}
\title{Retreives the different IDs of the available OpenCL platforms.}
\description{
   Retreives the different IDs of the available OpenCL platforms.
}
\details{
  Calls clGetPlatformIDs and returns a list of platform IDs which can be used later on.
}
\references{
\url{http://www.khronos.org/opencl/}
}
\keyword{OpenCL}