\name{getPlatformName}
\alias{getPlatformName}
\title{Retreives the name of the given OpenCL platform.}
\description{
   Retreives the name of the given OpenCL platform.
}
\details{
  Calls clGetPlatformInfo and returns a the name of the platform as a character string.
}
\references{
\url{http://www.khronos.org/opencl/}
}
\keyword{OpenCL}
