\name{ROpenCL-package}
\alias{ROpenCL-package}
\alias{ROpenCL}
\docType{package}
\title{
Access the OpenCL API from R
}
\description{
Parallel Computing for R Using OpenCL}
\details{
\tabular{ll}{
Package: \tab ROpenCL\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2011-06-02\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
Willem Ligtenberg

Maintainer: Willem Ligtenberg <willem.ligtenberg@openanalytics.eu>
}
\references{
  \url{http://www.khronos.org/opencl/}
}
\keyword{ package }
%\examples{
%% ~~ simple examples of the most important functions ~~
%}
