\name{setKernelArgInt}
\alias{setKernelArgInt}
\title{Sets a kernel argument with an integer value.}
\description{
   Sets a kernel argument with an integer value.
}
\details{
  Calls clSetKernelArg to set an integer kernel argument.
}
\references{
\url{http://www.khronos.org/opencl/}
}
\keyword{OpenCL}