\name{setKernelArgMem}
\alias{setKernelArgMem}
\title{Sets a kernel argument with a given buffer.}
\description{
   Sets a kernel argument with a given buffer.
}
\details{
  Calls clSetKernelArg to set a kernel argument using a given buffer object.
}
\references{
\url{http://www.khronos.org/opencl/}
}
\keyword{OpenCL}