\name{createBuffer}
\alias{createBuffer}
\title{Creates the right OpenCL buffer given an R Object}
\description{
   Creates the right OpenCL buffer given an R Object
}
\details{
  Depending on the input it calls different methods in ROpenCL
}
\references{
\url{http://www.khronos.org/opencl/}
}
\keyword{OpenCL}