\name{createProgram}
\alias{createProgram}
\title{Creates an OpenCL program.}
\description{
   Creates and OpenCL program given context, kernelSrc, kernelName, and parameters.
}
\details{
  High level wrapper to make creating an OpenCL program easier.
}
\references{
\url{http://www.khronos.org/opencl/}
}
\keyword{OpenCL}