\name{enqueueReadBuffer}
\alias{enqueueReadBuffer}
\title{Enqueues a read buffer}
\description{
   Enqueues a read buffer, depending on the class of the object.
}
\details{
  Calls the right enqueueReadBuffer function for the class of the object that was passed as the argument.
}
\references{
\url{http://www.khronos.org/opencl/}
}
\keyword{OpenCL}