\name{enqueueWriteBuffer}
\alias{enqueueWriteBuffer}
\title{Enqueues a write buffer}
\description{
   Enqueues a write buffer, depending on the class of the object.
}
\details{
  Calls the right enqueueWriteBuffer function for the class of the object that was passed as the argument.
}
\references{
\url{http://www.khronos.org/opencl/}
}
\keyword{OpenCL}