\name{ROpenCL}
\alias{ROpenCL}
\docType{package}
\title{
Provide an easy interface to OpenCL computing.
}
\description{
ROpenCL tries to bridge OpenCL and R. It allows you to use OpenCL from within R.
It also provides was to transparently send R objects to OpenCL and retreive R objects back.
}
\details{
\tabular{ll}{
Package: \tab ROpenCL\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2011-09-06\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Willem Ligtenberg

Maintainer: Willem Ligtenberg <willem.ligtenberg@openanalytics.eu>
}
\references{
\url{http://www.khronos.org/opencl/}
}
\keyword{ OpenCL, HPC }
